fix: atomically claim terminal cdr rows

+++ src/Core/Workers/Libs/WorkerCallEvents/UpdateDataInDB.php
@@ -51,9 +51,10 @@ class UpdateDataInDB
             return;
         }
         try {
+            $token = trim((string)($data['processing_token'] ?? ''));
             $filter = [
-                "UNIQUEID=:id:",
-                'bind' => ['id' => $data['UNIQUEID'],],
+                'UNIQUEID=:id: AND processing_token=:token: AND work_completed<>1',
+                'bind' => ['id' => $data['UNIQUEID'], 'token' => $token],
             ];
             /** @var CallDetailRecordsTmp $m_data */
             $m_data = CallDetailRecordsTmp::findFirst($filter);
@@ -77,7 +78,7 @@ class UpdateDataInDB
 
             self::sendUserEventData($m_data, $data);
 
-            if ($res && $m_data->work_completed === "1") {
+            if ($res && $m_data->work_completed === "1" && $m_data->wasMovedToGeneral()) {
                 // Delete data from the temporary table, as they have already been moved to the permanent one.
                 $m_data->delete();
             }
@@ -110,7 +111,9 @@ class UpdateDataInDB
                 $insert_data['to_account'],
                 $insert_data['appname'],
                 $insert_data['is_app'],
-                $insert_data['transfer']
+                $insert_data['transfer'],
+                $insert_data['processing_token'],
+                $insert_data['processing_started_at']
             );
             $am = Util::getAstManager('off');
             $am->UserEvent('CdrConnector', ['AgiData' => AsteriskManager::encodeCdrData($insert_data)]);
